Restructure files + namespaces
