+++ include/ui/screen/UserSelect.hpp
@@ -19,11 +19,11 @@ namespace Screen {
             Aether::List * list;
 
             // Stores copy of user vector
-            std::vector<User *> users;
+            std::vector<NX::User *> users;
 
         public:
             // Constructor takes display element and vector of User objects and sets up elements
-            UserSelect(Main::Application *, std::vector<User *>);
+            UserSelect(Main::Application *, std::vector<NX::User *>);
 
             // User images + names are handled + elements created on load
             void onLoad();
